pyproject.toml
[project]
name = "compliance-platform"
version = "1.0.0"
description = "Unified compliance platform for DPDP, SOC2, GDPR, HIPAA, ISO27001, NIST, CMMC"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Compliance Team", email = "team@compliance.example.com"}
]
dependencies = [
    "fastapi==0.109.2",
    "uvicorn[standard]==0.27.1",
    "pydantic==2.6.1",
    "pydantic-settings==2.1.0",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.1",
    "sqlalchemy[asyncio]==2.0.25",
    "alembic==1.13.1",
    "asyncpg==0.29.0",
    "redis==5.0.1",
    "celery[redis]==5.3.6",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.9",
    "httpx==0.26.0",
    "tenacity==8.2.3",
    "structlog==24.1.0",
    "python-json-logger==2.0.7",
    "prometheus-client==0.19.0",
    "openpyxl==3.1.2",
    "reportlab==4.1.0",
    "weasyprint==60.2",
    "jinja2==3.1.3",
    "aiofiles==23.2.1",
    "presidio-analyzer==2.2.351",
    "presidio-anonymizer==2.2.351",
    "spacy==3.7.4",
    "scikit-learn==1.4.1.post1",
    "numpy==1.26.4",
    "pandas==2.1.4",
]

[project.optional-dependencies]
dev = [
    "black==24.1.1",
    "isort==5.13.2",
    "ruff==0.2.2",
    "mypy==1.8.0",
    "pre-commit==3.6.2",
    "pytest==8.0.0",
    "pytest-asyncio==0.23.4",
    "pytest-cov==4.1.0",
    "httpx==0.26.0",
    "factory-boy==3.3.1",
    "faker==24.4.0",
    "testcontainers==3.1.0",
]
ml = [
    "torch==2.1.2",
    "torchvision==0.16.2",
    "transformers==4.36.2",
    "xgboost==2.0.3",
    "lightgbm==4.1.0",
    "mlflow==2.9.2",
    "optuna==3.5.0",
]
all = ["compliance-platform[dev,ml]"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM", "ARG"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[build-system]
requires = ["setuptools>=69.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
